# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenRisk API
  version: 1.0.0
  description: |
    Modern, intelligent risk management platform API.
    Part of the OpenDefender suite with AI-driven risk assessment, mitigation tracking, and security integration.
    
    **Authentication**: All protected endpoints require a Bearer JWT token in the `Authorization` header.
    
    **Base URL**: `/api/v1`
  contact:
    name: OpenDefender Team
  license:
    name: Business Source License 1.1 (BUSL-1.1)
    url: https://mariadb.com/bsl11/
servers:
  - url: /api/v1
    description: Production API
...
paths:
...
  /mitigations/{id}/subactions/{subactionId}:
    delete:
      tags:
        - Mitigations
      summary: Delete mitigation sub-action
      operationId: deleteSubAction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Mitigation UUID
        - name: subactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Sub-action UUID
      responses:
        '204':
          description: Sub-action deleted
        '401':
          description: Unauthorized
        '404':
          description: Sub-action not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token. Obtain via /auth/login
...